requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
asgiref==3.3.1
async-exit-stack==1.0.1
async-generator==1.10
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.2
Django==3.1.7
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.0.0
h11==0.12.0
httptools==0.1.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflect==5.3.0
itsdangerous==1.1.✉0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
orjson==3.5.1
passlib==1.7.4
Pillow==8.2.0
promise==2.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
pymitter==0.3.0
pyOpenSSL==20.0.1
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==1.6.1
service-identity==18.1.0
six==1.15.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.0
sqlparse==0.4.1
sse-starlette==0.6.2
starlette==0.13.6
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
zope.interface==5.4.0
